pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "futures-bot"
version = "1.0.0"
description = "Production-Ready Trading Signal System"
authors = [
    {name = "Trading System", email = "trading@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "optuna>=3.2.0",
    "joblib>=1.3.0",
    "ta>=0.10.2",
    "ta-lib>=0.4.25",
    "websockets>=11.0.3",
    "aiohttp>=3.8.5",
    "python-binance>=1.0.17",
    "requests>=2.31.0",
    "pyarrow>=12.0.0",
    "fastparquet>=2023.8.0",
    "pydantic>=2.3.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.4",
    "hvac>=1.2.1",
    "prometheus-client>=0.17.1",
    "structlog>=23.1.0",
    "rich>=13.5.2",
    "click>=8.1.7",
    "plotly>=5.15.0",
    "jinja2>=3.1.2",
    "weasyprint>=59.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.1",
    "pre-commit>=3.3.3",
]

[project.scripts]
futures-train = "scripts.train:main"
futures-serve = "scripts.serve:main"
futures-download = "scripts.download_data:main"
futures-backtest = "scripts.backtest:main"
futures-replay = "scripts.replay_historical:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*", "config*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "B", "C90", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]